fix: modify dateKeyToString function to dateAndBigIntToString
